can toggle side nav in mobile view
